pyproject.toml
[build-system]
requires = ["hatchling>=1.21"]
build-backend = "hatchling.build"

[project]
name = "agentvuln"
version = "0.1.0"
description = "Runtime guardrail SDK for agentic AI systems — blocks prompt injection, exfiltration, and unsafe tool calls before they happen."
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.10"
authors = [{ name = "Emmanuel GJR", email = "emmanuelgjr@gmail.com" }]
keywords = [
    "ai-security",
    "llm-security",
    "prompt-injection",
    "guardrails",
    "agent-security",
    "owasp-llm",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
]
dependencies = [
    "pydantic>=2.6,<3",
]

[project.optional-dependencies]
server = [
    "fastapi>=0.115,<1",
    "uvicorn>=0.30,<1",
    "jinja2>=3.1,<4",
    "python-multipart>=0.0.9",
    "httpx>=0.27,<1",
    "PyJWT[crypto]>=2.9,<3",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.6",
]

[project.urls]
Homepage = "https://github.com/emmanuelgjr/AgentVulnMitigator"
Repository = "https://github.com/emmanuelgjr/AgentVulnMitigator"
Issues = "https://github.com/emmanuelgjr/AgentVulnMitigator/issues"

[tool.hatch.build.targets.wheel]
packages = ["agentvuln"]

[tool.hatch.build.targets.sdist]
include = [
    "agentvuln/**",
    "README.md",
    "LICENSE",
    "pyproject.toml",
]

[tool.ruff]
line-length = 110
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "S", "RUF"]
ignore = ["S101"]  # allow asserts in tests

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S", "B", "E501", "RUF001", "RUF003"]
"agentvuln/core/normalize.py" = ["RUF001"]
"agentvuln/core/mitigation.py" = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
